refactor(client): split chat_input_bar.dart into focused widgets (#759)
